Desktop/pic-n-eat-v2/backend/requirements.txt
# Backend API Requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
cloudinary>=1.36.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
bcrypt>=4.1.0

# Cloud Storage (AWS S3)
boto3>=1.29.0
botocore>=1.32.0

# HTTP Requests
httpx>=0.25.0
requests>=2.31.0

# Image Processing
pillow>=10.0.0
python-magic>=0.4.27

# ML Inference
torch>=2.0.0
torchvision>=0.15.0
onnxruntime>=1.15.0

# Monitoring & Logging
prometheus-client>=0.18.0
python-json-logger>=2.0.7

# Validation
email-validator>=2.1.0
